# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface FoodEntry {
  id: number;
  entryId: string;
  date: string;
  name: string;
  brand?: string | null;
  servingG: number;
  caloriesPer100g: number;
  proteinPer100g: number;
  carbsPer100g: number;
  fatPer100g: number;
  fiberPer100g?: number | null;
  addedAt: number;
}

export interface CreateFoodEntry {
  entryId: string;
  date: string;
  name: string;
  brand?: string | null;
  servingG: number;
  caloriesPer100g: number;
  proteinPer100g: number;
  carbsPer100g: number;
  fatPer100g: number;
  fiberPer100g?: number | null;
  addedAt: number;
}

export type UserProfileSex = typeof UserProfileSex[keyof typeof UserProfileSex];


export const UserProfileSex = {
  male: 'male',
  female: 'female',
} as const;

export type UserProfileActivityLevel = typeof UserProfileActivityLevel[keyof typeof UserProfileActivityLevel];


export const UserProfileActivityLevel = {
  sedentary: 'sedentary',
  light: 'light',
  moderate: 'moderate',
  active: 'active',
  very_active: 'very_active',
} as const;

export type UserProfileGoal = typeof UserProfileGoal[keyof typeof UserProfileGoal];


export const UserProfileGoal = {
  lose: 'lose',
  maintain: 'maintain',
  gain: 'gain',
} as const;

export interface UserProfile {
  id: number;
  profileKey: string;
  sex: UserProfileSex;
  age: number;
  heightCm: number;
  weightKg: number;
  activityLevel: UserProfileActivityLevel;
  goal: UserProfileGoal;
}

export type UpsertProfileSex = typeof UpsertProfileSex[keyof typeof UpsertProfileSex];


export const UpsertProfileSex = {
  male: 'male',
  female: 'female',
} as const;

export type UpsertProfileActivityLevel = typeof UpsertProfileActivityLevel[keyof typeof UpsertProfileActivityLevel];


export const UpsertProfileActivityLevel = {
  sedentary: 'sedentary',
  light: 'light',
  moderate: 'moderate',
  active: 'active',
  very_active: 'very_active',
} as const;

export type UpsertProfileGoal = typeof UpsertProfileGoal[keyof typeof UpsertProfileGoal];


export const UpsertProfileGoal = {
  lose: 'lose',
  maintain: 'maintain',
  gain: 'gain',
} as const;

export interface UpsertProfile {
  sex: UpsertProfileSex;
  age: number;
  heightCm: number;
  weightKg: number;
  activityLevel: UpsertProfileActivityLevel;
  goal: UpsertProfileGoal;
}

export type GetEntriesParams = {
/**
 * Filter by date (YYYY-MM-DD)
 */
date?: string;
};