# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NutriScan API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: entries
    description: Food log entries
  - name: profile
    description: User profile

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /entries:
    get:
      operationId: getEntries
      tags: [entries]
      summary: Get food entries, optionally filtered by date
      parameters:
        - name: date
          in: query
          required: false
          description: Filter by date (YYYY-MM-DD)
          schema:
            type: string
      responses:
        "200":
          description: List of food entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodEntry"
    post:
      operationId: createEntry
      tags: [entries]
      summary: Add a food entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFoodEntry"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntry"

  /entries/{entryId}:
    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete a food entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: No profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfile"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]

    FoodEntry:
      type: object
      properties:
        id:
          type: integer
        entryId:
          type: string
        date:
          type: string
        name:
          type: string
        brand:
          type: string
          nullable: true
        servingG:
          type: integer
        caloriesPer100g:
          type: number
        proteinPer100g:
          type: number
        carbsPer100g:
          type: number
        fatPer100g:
          type: number
        fiberPer100g:
          type: number
          nullable: true
        addedAt:
          type: integer
      required:
        - id
        - entryId
        - date
        - name
        - servingG
        - caloriesPer100g
        - proteinPer100g
        - carbsPer100g
        - fatPer100g
        - addedAt

    CreateFoodEntry:
      type: object
      properties:
        entryId:
          type: string
        date:
          type: string
        name:
          type: string
        brand:
          type: string
          nullable: true
        servingG:
          type: integer
        caloriesPer100g:
          type: number
        proteinPer100g:
          type: number
        carbsPer100g:
          type: number
        fatPer100g:
          type: number
        fiberPer100g:
          type: number
          nullable: true
        addedAt:
          type: integer
      required:
        - entryId
        - date
        - name
        - servingG
        - caloriesPer100g
        - proteinPer100g
        - carbsPer100g
        - fatPer100g
        - addedAt

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        profileKey:
          type: string
        sex:
          type: string
          enum: [male, female]
        age:
          type: integer
        heightCm:
          type: integer
        weightKg:
          type: number
        activityLevel:
          type: string
          enum: [sedentary, light, moderate, active, very_active]
        goal:
          type: string
          enum: [lose, maintain, gain]
      required:
        - id
        - profileKey
        - sex
        - age
        - heightCm
        - weightKg
        - activityLevel
        - goal

    UpsertProfile:
      type: object
      properties:
        sex:
          type: string
          enum: [male, female]
        age:
          type: integer
        heightCm:
          type: integer
        weightKg:
          type: number
        activityLevel:
          type: string
          enum: [sedentary, light, moderate, active, very_active]
        goal:
          type: string
          enum: [lose, maintain, gain]
      required:
        - sex
        - age
        - heightCm
        - weightKg
        - activityLevel
        - goal