# lib/api-zod/src/generated/types/foodEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan API
 * OpenAPI spec version: 0.1.0
 */

export interface FoodEntry {
  id: number;
  entryId: string;
  date: string;
  name: string;
  brand?: string | null;
  servingG: number;
  caloriesPer100g: number;
  proteinPer100g: number;
  carbsPer100g: number;
  fatPer100g: number;
  fiberPer100g?: number | null;
  addedAt: number;
}