# lib/api-zod/src/generated/types/upsertProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan API
 * OpenAPI spec version: 0.1.0
 */
import type { UpsertProfileActivityLevel } from './upsertProfileActivityLevel';
import type { UpsertProfileGoal } from './upsertProfileGoal';
import type { UpsertProfileSex } from './upsertProfileSex';

export interface UpsertProfile {
  sex: UpsertProfileSex;
  age: number;
  heightCm: number;
  weightKg: number;
  activityLevel: UpsertProfileActivityLevel;
  goal: UpsertProfileGoal;
}