# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileActivityLevel } from './userProfileActivityLevel';
import type { UserProfileGoal } from './userProfileGoal';
import type { UserProfileSex } from './userProfileSex';

export interface UserProfile {
  id: number;
  profileKey: string;
  sex: UserProfileSex;
  age: number;
  heightCm: number;
  weightKg: number;
  activityLevel: UserProfileActivityLevel;
  goal: UserProfileGoal;
}